# dist/api.yml
openapi: 3.0.0
info:
  title: Sadna League API
  description: API for project
  version: 1.0.0
...
paths:
  /Register:
    post:
      tags:
        - Auth
      summary: Add new user
      description: A new user from the client is added to the server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: A new user has been added
        "400":
          description: wrong input parameters
...
  /judges/Register:
    post:
      tags:
        - Judges
      summary: Make an existing user a judge
      description: A logged in user will become a judge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Judge"
      responses:
        "201":
          description: Judge Created Successfully
        "400":
          description: wrong input parameters
...
components:
  schemas:
    User:
      type: object
      properties:
        username:
          maxLength: 8
          minLength: 3
          type: string
          description: username conatins between 3 to 8 letters only
          example: ladygaga
        firstname:
          type: string
          description: user's first name
          example: Stefani
        lastname:
          type: string
          description: user's last name
          example: Germanotta
        country:
          type: string
          description: user's country
          example: USA
        password:
          maxLength: 10
          minLength: 5
          type: string
          description:
            password conatins between 5 to 10 chars, at least one of them
            is a number and a special char
          format: password
          example: lady@56
        email:
          pattern: email
          type: string
          description: user's email
          example: ladygaga@gmail.com
        image-url:
          pattern: URL
          type: string
          description: URL for the user's image
          example: https://cloudinary.com/
      description: Model containing user info
...
    Judge:
      type: object
      properties:
        leagueID:
          type: integer
          description: League ID
          example: 1
        season:
          type: string
          description: current season name
          example: "2020/2021"
...